refactor: Simplifica pacote base do projeto para com.petdoc
