+++ src/main/java/com/petdoc/PetdocAppApplication.java
@@ -1,4 +1,4 @@
-package com.petdoc.backend;
+package com.petdoc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
