+++ src/main/java/com/petdoc/model/Pet.java
@@ -1,4 +1,4 @@
-package model;
+package com.petdoc.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
