+++ src/main/java/com/petdoc/model/Tutor.java
@@ -1,4 +1,4 @@
-package model;
+package com.petdoc.model;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
